+++ extensions/gumroad/CHANGELOG.md
@@ -1,3 +1,6 @@
 # Gumroad Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- View Products (basic and some details)
+- View Sales